backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

ollama>=0.3.0
pinecone-client>=3.0.0
langchain-text-splitters>=0.2.0
pika>=1.3.0
celery>=5.3.0
redis>=5.0.0
